fix: resolve remaining Kotlin compilation errors
- Use explicit types for ShizukuShell in ArchiveHandler
- Fix Log.w overload resolution by using (tag, tr) signature
- Use anonymous object for ComponentCallbacks2 in AppearanceUtils
- Make AudioMetadata public to fix visibility exposure
- Handle nullability in StorageUtils and TarUtils casts
- Fix type mismatches in BatchBackupOptions

+++ app/src/main/java/io/github/muntashirakon/AppManager/utils/TarUtils.kt
@@ -224,7 +224,7 @@ object TarUtils {
                             ExUtils.exceptionAsIgnored {
                                 Paths.setPermissions(
                                     file, finalEntry!!.mode,
-                                    finalEntry.longUserId, finalEntry.longGroupId
+                                    finalEntry.longUserId.toInt(), finalEntry.longGroupId.toInt()
                                 )
                             }
                             // Restore timestamp
